move types to types package
