+++ packages/action/src/utils.ts
@@ -1,6 +1,6 @@
 import fs from "fs";
 import type { FunctionPipe } from "../../types/src";
-import type { ConfigFile } from "@covector/files/src/schema";
+import type { ConfigFile } from "@covector/types";
 import type { GitHub } from "@actions/github/lib/utils";
 
 export const commandText = (pkg: {
